+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v1.1.0
 
 ### Added
+- Added `publicPlanbyDMPId` resolver, to query only public plans for the landing page [#293]
+- Added `getPlanVersions` and `extractPlanOutputs` for the `publicPlanByDMPId` resolver [#293]
 - Added override for `minimatch`
 - Added versionedTemplate schema and resolver so we can fetch a versioned template by its id.
 - Added `questionTags` and `versionedQuestionTags` tables [#274]